+++ ark/json-schema/index.ts
@@ -1,3 +1,3 @@
 export * from "./errors.ts"
-export { parseJsonSchema } from "./json.ts"
+export { jsonSchemaToType } from "./json.ts"
 export * from "./scope.ts"