+++ ark/type/__tests__/defaults.test.ts
@@ -938,4 +938,18 @@ contextualize(() => {
 			}).throws.snap("ParseError: Default foo must be a number (was a string)")
 		})
 	})
+
+	it("extracted from cyclic type", () => {
+		const T = type({
+			defaulted: "number = 0",
+			"nested?": "this"
+		})
+
+		const t = T.assert({})
+
+		attest(t).equals({ defaulted: 0 })
+
+		attest<number>(t.defaulted)
+		attest<number | undefined>(t.nested?.defaulted)
+	})
 })